+++ gfx/video_shader_parse.c
@@ -535,7 +535,7 @@ static struct video_shader_parameter *video_shader_parse_find_parameter(
       struct video_shader_parameter *params,
       unsigned num_params, const char *id)
 {
-   int i;
+   size_t i;
 
    for (i = 0; i < num_params; i++)
    {
@@ -557,7 +557,7 @@ static struct video_shader_parameter *video_shader_parse_find_parameter(
  **/
 void video_shader_resolve_parameters(struct video_shader *shader)
 {
-   int i;
+   size_t i;
    struct video_shader_parameter *param = &shader->parameters[0];
 
    shader->num_parameters = 0;
@@ -678,7 +678,7 @@ void video_shader_resolve_parameters(struct video_shader *shader)
 bool video_shader_load_current_parameter_values(
       config_file_t *conf, struct video_shader *shader)
 {
-   int i;
+   size_t i;
 
    if (!conf)
       return false;
@@ -781,7 +781,7 @@ static void shader_write_fbo(config_file_t *conf,
 static bool video_shader_write_root_preset(const struct video_shader *shader,
       const char *path)
 {
-   int i;
+   size_t i;
    char key[64];
    bool ret             = true;
    char *tmp            = (char*)malloc(3 * PATH_MAX_LENGTH);
@@ -1113,7 +1113,7 @@ static bool video_shader_write_referenced_preset(
       const char *path_to_save,
       const struct video_shader *shader)
 {
-   int i;
+   size_t i;
    config_file_t *conf                    = NULL;
    config_file_t *ref_conf                = NULL;
    struct video_shader *ref_shader        = (struct video_shader*)
@@ -1562,7 +1562,7 @@ static bool video_shader_load_root_config_into_shader(
       settings_t *settings,
       struct video_shader *shader)
 {
-   int i;
+   size_t i;
    unsigned num_passes = 0;
    bool watch_files    = settings->bools.video_shader_watch_files;
 
@@ -1674,7 +1674,7 @@ static bool video_shader_load_root_config_into_shader(
 static bool override_shader_values(config_file_t *override_conf,
       struct video_shader *shader)
 {
-   int i;
+   size_t i;
    bool return_val                     = false;
 
    if (!shader || !override_conf) 
@@ -1806,7 +1806,7 @@ static bool combine_shaders(struct video_shader *combined_shader,
                             struct video_shader *first_shader,
                             struct video_shader *second_shader)
 {
-   int i, j;
+   size_t i, j;
 
    for (i = 0; i < first_shader->passes && i <= GFX_MAX_SHADERS; i++)
    {
